Move code out of model class
