--- mlx/od/fcos/iou_loss.py
@@ -1,37 +1,0 @@
-import torch
-from torch import nn
-
-# Adapted from following to handle different ordering
-# https://github.com/tianzhi0549/FCOS/blob/master/maskrcnn_benchmark/layers/iou_loss.py
-class IOULoss(nn.Module):
-    def forward(self, pred, target, weight=None):
-        pred_left = pred[:, 1]
-        pred_top = pred[:, 0]
-        pred_right = pred[:, 3]
-        pred_bottom = pred[:, 2]
-
-        target_left = target[:, 1]
-        target_top = target[:, 0]
-        target_right = target[:, 3]
-        target_bottom = target[:, 2]
-
-        target_area = (target_left + target_right) * \
-                      (target_top + target_bottom)
-        pred_area = (pred_left + pred_right) * \
-                    (pred_top + pred_bottom)
-
-        w_intersect = torch.min(pred_left, target_left) + \
-                      torch.min(pred_right, target_right)
-        h_intersect = torch.min(pred_bottom, target_bottom) + \
-                      torch.min(pred_top, target_top)
-
-        area_intersect = w_intersect * h_intersect
-        area_union = target_area + pred_area - area_intersect
-
-        losses = -torch.log((area_intersect + 1.0) / (area_union + 1.0))
-
-        if weight is not None and weight.sum() > 0:
-            return (losses * weight).sum() / weight.sum()
-        else:
-            assert losses.numel() != 0
-            return losses.mean()
